We need cloud-platform scope in order to authenticate to the K8s master.
